some simplifications and renamings
