+++ packages/api/src/application/city/city.module.test.ts
@@ -1,7 +1,7 @@
 import { Injectable, Module } from '@nestjs/common';
 import { ChainableCityProvider } from 'src/domain/city/chainable-city.provider';
 import { CityProvider } from 'src/domain/city/city-provider.abstract';
-import { CityProviderAdapter } from 'src/infrastructure/city/adapters/city-provider.adapter';
+import { CityProviderAdapter } from 'src/infrastructure/city/wrappers/city-provider.adapter';
 import { OpenWeatherCityProvider } from '../../infrastructure/city/providers/openweather.provider';
 import { WeatherApiCityProvider } from '../../infrastructure/city/providers/weatherapi.provider';
 import { CityExistsConstraint } from './validators/city-exists.constraint';
