+++ packages/api/src/application/city/city.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { CityProvider } from 'src/domain/city/city-provider.abstract';
 import { CacheModule } from 'src/infrastructure/cache/cache.module';
 import { CacheAccessor } from 'src/infrastructure/cache/interfaces/cache-service.interface';
-import { CityProviderAdapter } from 'src/infrastructure/city/adapters/city-provider.adapter';
-import { CityProviderLoggingDecorator } from 'src/infrastructure/city/decorators/city-provider-logging.decorator';
-import { CityProviderCacheProxy } from 'src/infrastructure/city/proxies/city-validation-cache.proxy';
+import { CityProviderLoggingDecorator } from 'src/infrastructure/city/wrappers/city-provider-logging.decorator';
+import { CityProviderAdapter } from 'src/infrastructure/city/wrappers/city-provider.adapter';
+import { CityProviderCacheProxy } from 'src/infrastructure/city/wrappers/city-validation-cache.proxy';
 import { LoggerModule } from 'src/infrastructure/logger/logger.module';
 import { LoggerService } from 'src/infrastructure/logger/logger.service';
 import { OpenWeatherCityProvider } from '../../infrastructure/city/providers/openweather.provider';
