+++ packages/api/src/application/notifications/constants/notification.type.ts
@@ -1,0 +1,14 @@
+import { WeatherUpdate } from './weather-update.interface';
+
+export type Notification = {
+  to: string;
+  subject: string;
+};
+
+export type ConfirmationNotification = Notification & {
+  token: string;
+};
+
+export type WeatherUpdateNotification = Notification & {
+  data: WeatherUpdate;
+};
