--- packages/api/src/application/notifications/constants/types/notification.interface.ts
@@ -1,14 +1,0 @@
-import { WeatherUpdate } from '../../../../infrastructure/notifications/types/weather-update.interface';
-
-export interface Notification {
-  to: string;
-  subject: string;
-}
-
-export interface ConfirmationNotification extends Notification {
-  token: string;
-}
-
-export interface WeatherUpdateNotification extends Notification {
-  data: WeatherUpdate;
-}
