+++ packages/api/src/application/notifications/constants/weather-update.interface.ts
