+++ packages/api/src/application/notifications/services/notifications.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable, InternalServerErrorException } from '@nestjs/common';
-import { NotificationType } from 'src/application/notifications/constants/enums/notification-type.enum';
-import { ConfirmationNotification, WeatherUpdateNotification } from 'src/application/notifications/constants/types/notification.interface';
+import { NotificationType } from 'src/application/notifications/constants/notification-type.enum';
+import { ConfirmationNotification, WeatherUpdateNotification } from 'src/application/notifications/constants/notification.type';
 import { NotificationsSender, NotificationsSenderToken } from 'src/domain/notifications/notifications-sender.interface';
 import { NotificationsServiceInterface } from 'src/domain/notifications/notifications-service.abstract';
