+++ packages/api/src/application/subscriptions/dtos/create-subscription.dto.ts
@@ -1,6 +1,6 @@
 import { IsEmail, IsEnum, IsNotEmpty } from 'class-validator';
 import { CityExists } from 'src/application/city/validators/city-exists.validator';
-import { Frequency } from 'src/domain/subscription/valueObjects/frequency.vo';
+import { Frequency } from 'src/domain/subscription/frequency.vo';
 
 export class CreateSubscriptionDto {
   @IsEmail()
