+++ packages/api/src/application/weather/weather.module.test.ts
@@ -1,9 +1,9 @@
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { WeatherProvider } from 'src/domain/weather/interfaces/weather-provider.abstract';
-import { WeatherProviderAdapter } from 'src/infrastructure/weather/adapters/weather-povider.adapter';
 import { OpenWeatherWeatherProvider } from 'src/infrastructure/weather/providers/openweather.provider';
 import { WeatherApiWeatherProvider } from 'src/infrastructure/weather/providers/weatherapi.provider';
+import { WeatherProviderAdapter } from 'src/infrastructure/weather/wrappers/weather-povider.adapter';
 import { WeatherController } from 'src/presentation/weather.controller';
 
 @Module({
