+++ packages/api/src/application/weather/weather.module.ts
@@ -4,11 +4,11 @@ import { CacheModule } from 'src/infrastructure/cache/cache.module';
 import { CacheAccessor } from 'src/infrastructure/cache/interfaces/cache-service.interface';
 import { LoggerModule } from 'src/infrastructure/logger/logger.module';
 import { LoggerService } from 'src/infrastructure/logger/logger.service';
-import { WeatherProviderAdapter } from 'src/infrastructure/weather/adapters/weather-povider.adapter';
-import { WeatherProviderLoggingDecorator } from 'src/infrastructure/weather/decorators/weather-provider-logging.decorator';
 import { OpenWeatherWeatherProvider } from 'src/infrastructure/weather/providers/openweather.provider';
 import { WeatherApiWeatherProvider } from 'src/infrastructure/weather/providers/weatherapi.provider';
-import { WeatherProviderCacheProxy } from 'src/infrastructure/weather/proxies/weather-cache.proxy';
+import { WeatherProviderCacheProxy } from 'src/infrastructure/weather/wrappers/weather-cache.proxy';
+import { WeatherProviderAdapter } from 'src/infrastructure/weather/wrappers/weather-povider.adapter';
+import { WeatherProviderLoggingDecorator } from 'src/infrastructure/weather/wrappers/weather-provider-logging.decorator';
 import { WeatherController } from 'src/presentation/weather.controller';
 
 @Module({
