+++ packages/api/src/domain/notifications/notifications-service.abstract.ts
@@ -1,5 +1,5 @@
-import { NotificationType } from 'src/application/notifications/constants/enums/notification-type.enum';
-import { ConfirmationNotification, WeatherUpdateNotification } from 'src/application/notifications/constants/types/notification.interface';
+import { NotificationType } from 'src/application/notifications/constants/notification-type.enum';
+import { ConfirmationNotification, WeatherUpdateNotification } from 'src/application/notifications/constants/notification.type';
 
 export abstract class NotificationsServiceInterface {
   abstract sendConfirmationNotification(data: ConfirmationNotification, type: NotificationType): Promise<void>;
