+++ packages/api/src/infrastructure/city/providers/weatherapi.provider.ts
@@ -6,7 +6,7 @@ import { firstValueFrom } from 'rxjs';
 import { CityNotFoundException } from 'src/common/errors/city-not-found.error';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
 import { ChainableCityProvider } from 'src/domain/city/chainable-city.provider';
-import { WeatherApiCityFetch } from '../../shared/types/weatherapi-city-fetch.type';
+import { WeatherApiCityFetch } from '../../shared/constants/weatherapi-city-fetch.type';
 
 @Injectable()
 export class WeatherApiCityProvider extends ChainableCityProvider {
