+++ packages/api/src/infrastructure/notifications/mail/constants/email-notifications.type.ts
@@ -1,0 +1,9 @@
+import { ConfirmationNotification, WeatherUpdateNotification } from '../../../../application/notifications/constants/notification.type';
+
+export type EmailNotification = {
+  html?: string;
+};
+
+//empty yet, but possibly filled in future
+export type ConfirmationEmail = ConfirmationNotification & EmailNotification;
+export type UpdateEmail = WeatherUpdateNotification & EmailNotification;
