--- packages/api/src/infrastructure/notifications/mail/constants/types/email-notifications.interface.ts
@@ -1,9 +1,0 @@
-import { ConfirmationNotification, WeatherUpdateNotification } from '../../../../../application/notifications/constants/types/notification.interface';
-
-export interface EmailNotification {
-  html?: string;
-}
-
-//empty yet, but possibly filled in future
-export interface ConfirmationEmail extends ConfirmationNotification, EmailNotification {}
-export interface UpdateEmail extends WeatherUpdateNotification, EmailNotification {}
