+++ packages/api/src/infrastructure/scheduler/weather-scheduler/types/weather-update.options.ts
@@ -1,5 +1,5 @@
-import { Frequency } from 'src/domain/subscription/valueObjects/frequency.vo';
-import { NotificationSubjects } from '../../../../application/notifications/constants/enums/notification-subjects.enum';
+import { Frequency } from 'src/domain/subscription/frequency.vo';
+import { NotificationSubjects } from '../../../../application/notifications/constants/notification-subjects.enum';
 
 export type WeatherUpdateOptions = {
   frequency: Frequency;
