+++ packages/api/src/infrastructure/shared/constants/openweather-weather-fetch.type.ts
