+++ packages/api/src/infrastructure/shared/constants/weatherapi-city-fetch.type.ts
