+++ packages/api/src/infrastructure/shared/constants/weatherapi-weather-fetch.type.ts
@@ -1,4 +1,4 @@
-import { WeatherApiCityFetch } from 'src/infrastructure/shared/types/weatherapi-city-fetch.type';
+import { WeatherApiCityFetch } from 'src/infrastructure/shared/constants/weatherapi-city-fetch.type';
 
 export type WeatherApiWeatherFetch = {
   location: WeatherApiCityFetch;
