+++ packages/api/src/infrastructure/weather/mappers/openweather.mapper.ts
@@ -1,5 +1,5 @@
 import { Weather } from 'src/domain/weather/weather.entity';
-import { OpenWeatherWeatherFetch } from 'src/infrastructure/shared/types/openweather-weather-fetch.type';
+import { OpenWeatherWeatherFetch } from 'src/infrastructure/shared/constants/openweather-weather-fetch.type';
 
 export class OpenWeatherDtoMapper {
   static toEntity(data: OpenWeatherWeatherFetch): Weather {
