+++ packages/api/src/infrastructure/weather/mappers/weatherapi.mapper.ts
@@ -1,5 +1,5 @@
 import { Weather } from 'src/domain/weather/weather.entity';
-import { WeatherApiWeatherFetch } from 'src/infrastructure/shared/types/weatherapi-weather-fetch.type';
+import { WeatherApiWeatherFetch } from 'src/infrastructure/shared/constants/weatherapi-weather-fetch.type';
 
 export class WeatherApiDtoMapper {
   static toEntity(data: WeatherApiWeatherFetch): Weather {
