+++ packages/api/src/infrastructure/weather/providers/openweather.provider.ts
@@ -7,7 +7,7 @@ import { CityNotFoundException } from 'src/common/errors/city-not-found.error';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
 import { ChainableWeatherProvider } from 'src/domain/weather/interfaces/chainable-weather-provider.abstract';
 import { Weather } from 'src/domain/weather/weather.entity';
-import { OpenWeatherWeatherFetch } from 'src/infrastructure/shared/types/openweather-weather-fetch.type';
+import { OpenWeatherWeatherFetch } from 'src/infrastructure/shared/constants/openweather-weather-fetch.type';
 import { OpenWeatherDtoMapper } from '../mappers/openweather.mapper';
 
 @Injectable()
