+++ packages/api/src/infrastructure/weather/providers/weatherapi.provider.ts
@@ -7,7 +7,7 @@ import { CityNotFoundException } from 'src/common/errors/city-not-found.error';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
 import { ChainableWeatherProvider } from 'src/domain/weather/interfaces/chainable-weather-provider.abstract';
 import { Weather } from 'src/domain/weather/weather.entity';
-import { WeatherApiWeatherFetch } from 'src/infrastructure/shared/types/weatherapi-weather-fetch.type';
+import { WeatherApiWeatherFetch } from 'src/infrastructure/shared/constants/weatherapi-weather-fetch.type';
 import { WeatherApiDtoMapper } from '../mappers/weatherapi.mapper';
 
 @Injectable()
