+++ packages/api/test/integration/subscription.integration.spec.ts
@@ -5,8 +5,8 @@ import { Test } from '@nestjs/testing';
 import { useContainer } from 'class-validator';
 import { Model, Types } from 'mongoose';
 import { CityTestModule } from 'src/application/city/city.module.test';
-import { NotificationSubjects } from 'src/application/notifications/constants/enums/notification-subjects.enum';
-import { NotificationType } from 'src/application/notifications/constants/enums/notification-type.enum';
+import { NotificationSubjects } from 'src/application/notifications/constants/notification-subjects.enum';
+import { NotificationType } from 'src/application/notifications/constants/notification-type.enum';
 import { NotificationsTestModule } from 'src/application/notifications/notifications.module.test';
 import { CreateSubscriptionDto } from 'src/application/subscriptions/dtos/create-subscription.dto';
 import { SubscriptionTestModule } from 'src/application/subscriptions/subscriptions.module.test';
@@ -14,7 +14,7 @@ import { ExternalApiException } from 'src/common/errors/external-api.error';
 import { DatabaseExceptionFilter } from 'src/common/filters/database-exception.filter';
 import { appTestConfig, databaseTestConfig } from 'src/config/test.config';
 import { NotificationsServiceInterface } from 'src/domain/notifications/notifications-service.abstract';
-import { Frequency } from 'src/domain/subscription/valueObjects/frequency.vo';
+import { Frequency } from 'src/domain/subscription/frequency.vo';
 import { OpenWeatherCityProvider } from 'src/infrastructure/city/providers/openweather.provider';
 import { WeatherApiCityProvider } from 'src/infrastructure/city/providers/weatherapi.provider';
 import { DatabaseTestModule } from 'src/infrastructure/database/test/database.module.test';
